Adjust interface for compatibility with android lollipop 5.1
There are some changes in the android telephony framework between lollipop 5.0 and lollipop 5.1,
the parameter in some interfaces are "long" for android lollipop 5.0, but "int" for android lollipop 5.1.

The apk for the application layer was developed under android lollipop 5.0, so some interfaces in the framework are overloaded,
and the apks for the application layer use reflection to call the interface in the framework.